refactor: move fn to model file
